[Refactor] feature flag to implementation package (#95)
* Migration to "implementation" package for FeatureFlags

* package updates
