+++ azure-communication-ui/azure-communication-ui-demo-app/src/main/java/com/azure/android/communication/ui/callingcompositedemoapp/CallLauncherActivity.kt
@@ -15,7 +15,7 @@ import com.azure.android.communication.ui.callingcompositedemoapp.features.Addit
 import com.azure.android.communication.ui.callingcompositedemoapp.features.conditionallyRegisterDiagnostics
 import com.azure.android.communication.ui.callingcompositedemoapp.launcher.CallingCompositeLauncher
 import com.azure.android.communication.ui.callingcompositedemoapp.launcher.FeatureFlagView
-import com.azure.android.communication.ui.utilities.FeatureFlags
+import com.azure.android.communication.ui.utilities.implementation.FeatureFlags
 import java.util.UUID
 
 class CallLauncherActivity : AppCompatActivity() {
