+++ azure-communication-ui/azure-communication-ui-demo-app/src/main/java/com/azure/android/communication/ui/callingcompositedemoapp/launcher/FeatureFlagView.kt
@@ -5,8 +5,8 @@ import android.content.SharedPreferences
 import android.util.AttributeSet
 import android.widget.CheckBox
 import android.widget.LinearLayout
-import com.azure.android.communication.ui.utilities.FEATURE_FLAG_SHARED_PREFS_KEY
-import com.azure.android.communication.ui.utilities.FeatureFlags
+import com.azure.android.communication.ui.utilities.implementation.FEATURE_FLAG_SHARED_PREFS_KEY
+import com.azure.android.communication.ui.utilities.implementation.FeatureFlags
 
 // This lists all the Features in the FeatureFlag system
 // and lets you enable/disable them.
