+++ azure-communication-ui/azure-communication-ui/src/main/java/com/azure/android/communication/ui/utilities/implementation/FeatureFlags.kt
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.ui.utilities
+package com.azure.android.communication.ui.utilities.implementation
 
 import android.app.Application
 import android.content.Context
